Move over attention-bank tests
