+++ tests/cython/atomspace/test_atomspace.py
@@ -102,30 +102,6 @@ class AtomSpaceTest(TestCase):
         self.assertEqual(new_tv.mean, 0.75)
         self.assertAlmostEqual(new_tv.confidence, 0.9, places=4)
 
-    def test_attention_value(self):
-        node = Node("test")
-
-        # check values come back as assigned
-        node.sti = 1
-        node.lti = 2
-        node.vlti = 3
-        assert node.sti == 1
-        assert node.lti == 2
-        assert node.vlti == 3
-
-        # Check increment and decrement for vlti
-        node.decrement_vlti()
-        assert node.vlti == 2
-        node.increment_vlti()
-        assert node.vlti == 3
-
-        # Check dictionary setting and getting of av property.
-        node.av = {"sti": 4, "lti": 5, "vlti": 6}
-        assert node.sti == 4
-        assert node.lti == 5
-        assert node.vlti == 6
-        assert node.av == {"sti": 4, "lti": 5, "vlti": 6}
-
     def test_get_by_type(self):
         a1 = Node("test1")
         a2 = ConceptNode("test2")
@@ -151,31 +127,6 @@ class AtomSpaceTest(TestCase):
         # test empty
         result = self.space.get_atoms_by_type(types.AnchorNode, subtype=False)
         self.assertEqual(len(result), 0)
-
-    def test_get_by_av(self):
-        a1 = ConceptNode("test1")
-        a2 = ConceptNode("test2")
-        a3 = InheritanceLink(a1, a2)
-        a4 = ConceptNode("test4")
-        a5 = ConceptNode("test5")
-
-        a1.sti = 10
-        a2.sti = 5
-        a3.sti = 4
-        a4.sti = 1
-
-        #ImportanceIndex is Asynchronus give it some time
-        sleep(1)
-
-        result = self.space.get_atoms_by_av(4, 10)
-        print ("The atoms-by-av result is ", result)
-        assert len(result) == 3
-        assert set(result) == set([a1, a2, a3])
-        assert a4 not in result
-
-        result = self.space.get_atoms_in_attentional_focus()
-        assert len(result) == 4
-        assert set(result) == set([a1, a2, a3, a4])
 
     def test_incoming_by_type(self):
         a1 = Node("test1")
